dbStuff moved to database folder
