+++ database/models/departure.ts
@@ -5,7 +5,7 @@ import {
   InferCreationAttributes,
   CreationOptional,
 } from "@sequelize/core";
-import db from "../../backend/util/db";
+import db from "../db";
 
 import Line from "./line";
