+++ database/models/dock.ts
@@ -6,7 +6,7 @@ import {
   CreationOptional,
 } from "@sequelize/core";
 
-import db from "../../backend/util/db";
+import db from "../db";
 
 const { sequelize } = db;
