+++ database/models/line.ts
@@ -5,7 +5,7 @@ import {
   InferCreationAttributes,
   CreationOptional,
 } from "@sequelize/core";
-import db from "../../backend/util/db";
+import db from "../db";
 
 import Dock from "./dock";
